Import lockfile generator from personal dotfiles
The code in this repository is mostly inspired by a trick I used in my
dotfiles framework. After switching to NixOS, a lot of my dotfiles
framework became irrelevant. I need to rewrite it, which is part of what
this repository is for.

Anyway, I've copied over several files and modified them to work
standalone. Now I can use plugins from this repo as a custom nix
channel.