Remove unnecessary file type detector
An autocommand was supposedly setting the file type to `terminal` when
I opened new terminal windows. Maybe something changed, or maybe it
never worked to begin with, but neovim does this automatically.

Either way, it was silently failing my tests because vim 8 doesn't
recognize `TermOpen` as a valid autocmd event.

That was miserable to track down.