Return misc.vim & git.vim to dotfiles
This nursery experiment was useful so long as I was using nix channels.
I've been on flakes for a good while but too lazy to switch it back.
Well, I just did it, and I killed off a good chunk of dead code in the
process. Now this just repackages other repos. I'd like to remove that
too.